Divided statistic tests to their own files and updated deleting surveys
