+++ src/tests/services/survey_service_test.py
@@ -291,194 +291,6 @@ class TestSurveyService(unittest.TestCase):
             question_id)
         self.assertEqual(check, return_value)
 
-    def test_get_users_who_answered_survey(self):
-        repo_return_value = [1, "jorma@uotinen.com", "This is a group"]
-        self.repo_mock.get_users_who_answered_survey.return_value = repo_return_value
-        survey_id = 1
-
-        returned_value = self.survey_service.get_users_who_answered_survey(
-            survey_id)
-        self.repo_mock.get_users_who_answered_survey.assert_called_with(
-            survey_id)
-        self.assertEqual(returned_value, repo_return_value)
-
-    def test_get_number_of_submissions_for_survey_calls_repo_correctly(self):
-        self.repo_mock.get_number_of_submissions.return_value = 2
-        survey_id = 1
-        self.survey_service.get_number_of_submissions_for_survey(survey_id)
-        self.repo_mock.get_number_of_submissions.assert_called_with(
-            survey_id
-        )
-
-    def test_get_answer_distribution_for_survey_questions_calls_repo_correctly(self):
-        Row = namedtuple("Row", "q_id q_text a_id a_text count")
-        self.repo_mock.get_answer_distribution.return_value = [Row(1, "What is love?", 3,
-                                                                   "Baby don't hurt me", 2)]
-        survey_id = 1
-        self.survey_service.get_answer_distribution_for_survey_questions(
-            survey_id)
-        self.repo_mock.get_answer_distribution.assert_called_with(
-            survey_id, None, None, None, '')
-
-    def test_answer_distribution_of_unanswered_survey(self):
-        Row = namedtuple("Row", "q_id q_text a_id a_text count")
-        self.repo_mock.get_answer_distribution.return_value = [
-            Row(1, "What is love?", 3, "Baby don't hurt me", 0),
-            Row(2, "Do they know it's Christmas?", 5, "No", 0)]
-        result = self.survey_service.get_answer_distribution_for_survey_questions(
-            1)
-        self.assertIsNone(result)
-
-    def test_answer_distribution_no_submissions_match_filters(self):
-        self.repo_mock.get_answer_distribution.return_value = None
-        result = self.survey_service.get_answer_distribution_for_survey_questions(
-            2)
-        self.assertIsNone(result)
-
-    def test_answer_distribution_of_survey_with_submissions(self):
-        Row = namedtuple("Row", "q_id q_text a_id a_text count")
-        self.repo_mock.get_answer_distribution.return_value = [
-            Row(1, "What is love?", 3, "Baby don't hurt me", 0),
-            Row(2, "Do they know it's Christmas?", 5, "No", 3)]
-        result = self.survey_service.get_answer_distribution_for_survey_questions(
-            1)
-        self.assertTrue(result)
-
-    def test_get_users_who_answered_survey_in_timerange_returns_none_with_invalid_timerage(self):
-        start_date_1 = datetime.fromisoformat("2011-11-04 00:05:23.283")
-        end_date_1 = datetime.fromisoformat("2010-11-04 00:05:23.283")
-        start_date_2 = datetime.fromisoformat("2015-11-04 00:05:23.283")
-        end_date_2 = datetime.fromisoformat("2014-11-04 00:05:23.283")
-
-        service_response_1 = self.survey_service.get_users_who_answered_survey_in_timerange(
-            1, start_date_1, end_date_1)
-        service_response_2 = self.survey_service.get_users_who_answered_survey_in_timerange(
-            1, start_date_2, end_date_2)
-        self.assertIsNone(service_response_1)
-        self.assertIsNone(service_response_2)
-        self.repo_mock.get_users_who_answered_survey.assert_not_called()
-
-    def test_get_users_who_answered_survey_in_timerange_calls_repo_correctly_with_valid_timerange(self):
-        survey_id = 1
-        start_date = datetime.fromisoformat("2020-11-04 00:05:23.283")
-        end_date = datetime.fromisoformat("2021-11-04 00:05:23.283")
-        repo_value_to_return = [5, "timppa@gmail.com",
-                                "Boss Team", "2021-10-04 00:05:23.283"]
-
-        self.repo_mock.get_users_who_answered_survey.return_value = repo_value_to_return
-
-        survey_reponse = self.survey_service.get_users_who_answered_survey_in_timerange(
-            survey_id, start_date, end_date)
-
-        self.repo_mock.get_users_who_answered_survey.assert_called_once_with(
-            survey_id, start_date, end_date)
-        self.assertEqual(repo_value_to_return, survey_reponse)
-
-    def test_get_users_who_answered_survey_in_group_calls_repo_correctly_with_valid_group(self):
-        survey_id = 1
-        repo_value_to_return = [5, "timppa@gmail.com",
-                                "Boss Team", "2021-10-04 00:05:23.283"]
-        self.repo_mock.get_users_who_answered_survey.return_value = repo_value_to_return
-
-        survey_reponse = self.survey_service.get_users_who_answered_survey_in_group(
-            survey_id, group_name="Boss Team")
-
-        self.repo_mock.get_users_who_answered_survey.assert_called_once_with(
-            survey_id, group_name="Boss Team")
-        self.assertEqual(repo_value_to_return, survey_reponse)
-
-    def test_get_users_who_answered_survey_filtered_calls_repo_correctly_with_valid_arguments(self):
-        survey_id = 1
-        start_date = datetime.fromisoformat("2020-11-04 00:05:23.283")
-        end_date = datetime.fromisoformat("2021-11-04 00:05:23.283")
-
-        repo_value_to_return = [5, "timppa@gmail.com",
-                                "Boss Team", "2021-10-04 00:05:23.283"]
-        self.repo_mock.get_users_who_answered_survey.return_value = repo_value_to_return
-
-        survey_reponse = self.survey_service.get_users_who_answered_survey_filtered(
-            survey_id, start_date, end_date, "Boss Team", "boss@mail")
-
-        self.repo_mock.get_users_who_answered_survey.assert_called_once_with(
-            survey_id, start_date, end_date, "Boss Team", "boss@mail")
-        self.assertEqual(repo_value_to_return, survey_reponse)
-
-    def test_get_users_who_answered_survey_filtered_returns_none_with_invalid_timerage(self):
-        start_date_1 = datetime.fromisoformat("2011-11-04 00:05:23.283")
-        end_date_1 = datetime.fromisoformat("2010-11-04 00:05:23.283")
-
-        service_response_1 = self.survey_service.get_users_who_answered_survey_filtered(
-            1, start_date_1, end_date_1, "group", "mail")
-        self.assertIsNone(service_response_1)
-
-        self.repo_mock.get_users_who_answered_survey.assert_not_called()
-
-    def test_get_users_who_answered_survey_filtered_works_with_no_group(self):
-        survey_id = 1
-        start_date = datetime.fromisoformat("2020-11-04 00:05:23.283")
-        end_date = datetime.fromisoformat("2021-11-04 00:05:23.283")
-
-        repo_value_to_return = [5, "timppa@gmail.com",
-                                "", "2021-10-04 00:05:23.283"]
-        self.repo_mock.get_users_who_answered_survey.return_value = repo_value_to_return
-
-        survey_reponse = self.survey_service.get_users_who_answered_survey_filtered(
-            survey_id, start_date, end_date, None, "")
-
-        self.repo_mock.get_users_who_answered_survey.assert_called_once_with(
-            survey_id, start_date, end_date, None, "")
-        self.assertEqual(repo_value_to_return, survey_reponse)
-
-    def test_calculate_average_scores_by_category_calls_repo_correctly(self):
-        value_to_return = [(1, "Category 1", 1.0)]
-        self.repo_mock.calculate_average_scores_by_category.return_value = value_to_return
-        survey_id = 1
-        self.survey_service.calculate_average_scores_by_category(survey_id)
-        self.repo_mock.calculate_average_scores_by_category.assert_called_with(
-            survey_id,
-            None,
-            None,
-            None,
-            ""
-        )
-
-    def test_calculate_average_scores_by_category_does_not_call_repo_with_invalid(self):
-        survey_id = 1
-        start_date = datetime.fromisoformat("2020-11-04 00:05:23.283")
-        end_date = datetime.fromisoformat("2019-11-04 00:05:23.283")
-
-        return_value = self.survey_service.calculate_average_scores_by_category(
-            survey_id, start_date=start_date, end_date=end_date)
-        self.repo_mock.calculate_average_scores_by_category.assert_not_called()
-        self.assertIsNone(return_value)
-
-    def test_calculate_average_scores_by_category_calls_repo_correctly_when_filtering(self):
-        start_date = datetime.fromisoformat("2020-11-04 00:05:23.283")
-        end_date = datetime.fromisoformat("2021-11-04 00:05:23.283")
-        user_group_id = 22
-        user_group_name = "Group 1"
-        email = "mail.com"
-
-        survey_id = 1
-        repo_value_to_return = [(1, "Category 1", 1.0)]
-        service_value_to_return = [(1, "Category 1", 1.0, 1.0)]
-        self.repo_mock.calculate_average_scores_by_category.return_value = repo_value_to_return
-        returned_value = self.survey_service.calculate_average_scores_by_category(
-            survey_id,
-            user_group_name,
-            start_date,
-            end_date,
-            email
-        )
-        self.repo_mock.calculate_average_scores_by_category.assert_called_with(
-            survey_id,
-            user_group_name,
-            start_date,
-            end_date,
-            email
-        )
-        self.assertEqual(service_value_to_return, returned_value)
-
     def test_create_category_result_calls_repo_correctly(self):
         self.repo_mock.create_category_result.return_value = 1
         category_id = 1
